// api/seller/client/models/Campaign.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

/**
 * Model Campaign
 */
export type Campaign = {
  updatedAt: string;
  createdAt: string;
  userUpdated: string | null;
  userCreated: string | null;
  status: boolean;
  duaration: string | null;
  startDate: string;
  label: string;
  id: string;
};